+++ src/tkc/named_value.c
@@ -92,12 +92,7 @@ ret_t named_value_set_value(named_value_t* nv, const value_t* value) {
     return RET_OK;
   }
 
-  if (!value_equal(value, &(nv->value))) {
-    value_reset(&(nv->value));
-    return value_deep_copy(&(nv->value), value);
-  }
-
-  return RET_OK;
+  return value_replace(&(nv->value), value, TRUE);
 }
 
 ret_t named_value_deinit(named_value_t* nv) {